// wow_retail/wow_api_typescript-node/model/arrayRacesIndexData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { IndexData } from './indexData';

/**
* Array of races index
*/
export class ArrayRacesIndexData {
    'races': Array<IndexData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "races",
            "baseName": "races",
            "type": "Array<IndexData>"
        }    ];

    static getAttributeTypeMap() {
        return ArrayRacesIndexData.attributeTypeMap;
    }
}