// wow_retail/wow_api_typescript-node/model/arraySpecializationsIndexData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { IndexData } from './indexData';

/**
* Array of specializations index
*/
export class ArraySpecializationsIndexData {
    'characterSpecializations': Array<IndexData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "characterSpecializations",
            "baseName": "character_specializations",
            "type": "Array<IndexData>"
        }    ];

    static getAttributeTypeMap() {
        return ArraySpecializationsIndexData.attributeTypeMap;
    }
}