// wow_retail/wow_api_typescript-node/model/characterMediaData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AssetData } from './assetData';
import { CharacterMediaIndex } from './characterMediaIndex';

/**
* Character media data
*/
export class CharacterMediaData {
    'character': CharacterMediaIndex;
    'avatarUrl'?: string;
    'bustUrl'?: string;
    'renderUrl'?: string;
    'assets'?: Array<AssetData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "character",
            "baseName": "character",
            "type": "CharacterMediaIndex"
        },
        {
            "name": "avatarUrl",
            "baseName": "avatar_url",
            "type": "string"
        },
        {
            "name": "bustUrl",
            "baseName": "bust_url",
            "type": "string"
        },
        {
            "name": "renderUrl",
            "baseName": "render_url",
            "type": "string"
        },
        {
            "name": "assets",
            "baseName": "assets",
            "type": "Array<AssetData>"
        }    ];

    static getAttributeTypeMap() {
        return CharacterMediaData.attributeTypeMap;
    }
}