// wow_retail/wow_api_typescript-node/model/characterMediaIndex.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { RealmIndexData } from './realmIndexData';

/**
* Character media data
*/
export class CharacterMediaIndex {
    'name': string;
    'id': number;
    'realm': RealmIndexData;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "realm",
            "baseName": "realm",
            "type": "RealmIndexData"
        }    ];

    static getAttributeTypeMap() {
        return CharacterMediaIndex.attributeTypeMap;
    }
}