// wow_retail/wow_api_typescript-node/model/covenantProgressData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { IndexData } from './indexData';

/**
* Covenant progress data
*/
export class CovenantProgressData {
    'chosenCovenant': IndexData;
    'renownLevel': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "chosenCovenant",
            "baseName": "chosen_covenant",
            "type": "IndexData"
        },
        {
            "name": "renownLevel",
            "baseName": "renown_level",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CovenantProgressData.attributeTypeMap;
    }
}