// wow_retail/wow_api_typescript-node/model/genderData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { NameData } from './nameData';

/**
* Gender data
*/
export class GenderData {
    'male': NameData;
    'female': NameData;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "male",
            "baseName": "male",
            "type": "NameData"
        },
        {
            "name": "female",
            "baseName": "female",
            "type": "NameData"
        }    ];

    static getAttributeTypeMap() {
        return GenderData.attributeTypeMap;
    }
}