// wow_retail/wow_api_typescript-node/model/nameData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Different translations
*/
export class NameData {
    /**
    * The locale of English (U.S.)
    */
    'enUS': string;
    /**
    * The locale of mexican Spanish (Mexican)
    */
    'esMX': string;
    /**
    * The locale of Portuguese (Brazilian)
    */
    'ptBR': string;
    /**
    * The locale of German
    */
    'deDE': string;
    /**
    * The locale of English (U.K.)
    */
    'enGB': string;
    /**
    * The locale of Spanish (Modern)
    */
    'esES': string;
    /**
    * The locale of French
    */
    'frFR': string;
    /**
    * The locale of Italian
    */
    'itIT': string;
    /**
    * The locale of Russian
    */
    'ruRU': string;
    /**
    * The locale of Korean
    */
    'koKR': string;
    /**
    * The locale of Chinese (Taiwan/Traditional)
    */
    'zhTW': string;
    /**
    * The locale of Chinese (Simplified)
    */
    'zhCN': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "enUS",
            "baseName": "en_US",
            "type": "string"
        },
        {
            "name": "esMX",
            "baseName": "es_MX",
            "type": "string"
        },
        {
            "name": "ptBR",
            "baseName": "pt_BR",
            "type": "string"
        },
        {
            "name": "deDE",
            "baseName": "de_DE",
            "type": "string"
        },
        {
            "name": "enGB",
            "baseName": "en_GB",
            "type": "string"
        },
        {
            "name": "esES",
            "baseName": "es_ES",
            "type": "string"
        },
        {
            "name": "frFR",
            "baseName": "fr_FR",
            "type": "string"
        },
        {
            "name": "itIT",
            "baseName": "it_IT",
            "type": "string"
        },
        {
            "name": "ruRU",
            "baseName": "ru_RU",
            "type": "string"
        },
        {
            "name": "koKR",
            "baseName": "ko_KR",
            "type": "string"
        },
        {
            "name": "zhTW",
            "baseName": "zh_TW",
            "type": "string"
        },
        {
            "name": "zhCN",
            "baseName": "zh_CN",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return NameData.attributeTypeMap;
    }
}