// wow_retail/wow_api_typescript-node/model/playableClassData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GenderData } from './genderData';
import { IndexData } from './indexData';
import { NameData } from './nameData';

/**
* Playable class data
*/
export class PlayableClassData {
    /**
    * Id of the class
    */
    'id': number;
    'name': NameData;
    'genderName': GenderData;
    'powerType': IndexData;
    'specializations': Array<IndexData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "NameData"
        },
        {
            "name": "genderName",
            "baseName": "gender_name",
            "type": "GenderData"
        },
        {
            "name": "powerType",
            "baseName": "power_type",
            "type": "IndexData"
        },
        {
            "name": "specializations",
            "baseName": "specializations",
            "type": "Array<IndexData>"
        }    ];

    static getAttributeTypeMap() {
        return PlayableClassData.attributeTypeMap;
    }
}