// wow_retail/wow_api_typescript-node/model/playableRaceData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GenderData } from './genderData';
import { NameData } from './nameData';
import { TypeData } from './typeData';

/**
* Playable race data
*/
export class PlayableRaceData {
    /**
    * Id of the race
    */
    'id': number;
    'name': NameData;
    'genderName': GenderData;
    'faction': TypeData;
    'isSelectable': boolean;
    'isAlliedRace': boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "NameData"
        },
        {
            "name": "genderName",
            "baseName": "gender_name",
            "type": "GenderData"
        },
        {
            "name": "faction",
            "baseName": "faction",
            "type": "TypeData"
        },
        {
            "name": "isSelectable",
            "baseName": "is_selectable",
            "type": "boolean"
        },
        {
            "name": "isAlliedRace",
            "baseName": "is_allied_race",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return PlayableRaceData.attributeTypeMap;
    }
}