// wow_retail/wow_api_typescript-node/model/playableSpecializationData.ts
/**
 * Swagger BLIZZARD Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.6
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GenderData } from './genderData';
import { IndexData } from './indexData';
import { NameData } from './nameData';
import { PveTalentTiersData } from './pveTalentTiersData';
import { PvpTalentTiersData } from './pvpTalentTiersData';
import { TypeData } from './typeData';

/**
* Playable specilization data
*/
export class PlayableSpecializationData {
    /**
    * Id of the specilization
    */
    'id'?: number;
    'playableClass'?: IndexData;
    'name'?: NameData;
    'genderDescription'?: GenderData;
    'role'?: TypeData;
    'talentTiers'?: Array<PveTalentTiersData>;
    'pvpTalents'?: Array<PvpTalentTiersData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "playableClass",
            "baseName": "playable_class",
            "type": "IndexData"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "NameData"
        },
        {
            "name": "genderDescription",
            "baseName": "gender_description",
            "type": "GenderData"
        },
        {
            "name": "role",
            "baseName": "role",
            "type": "TypeData"
        },
        {
            "name": "talentTiers",
            "baseName": "talent_tiers",
            "type": "Array<PveTalentTiersData>"
        },
        {
            "name": "pvpTalents",
            "baseName": "pvp_talents",
            "type": "Array<PvpTalentTiersData>"
        }    ];

    static getAttributeTypeMap() {
        return PlayableSpecializationData.attributeTypeMap;
    }
}